pestilence/requirements/base.txt
# Containers and deployment
docker-compose==1.8.1

# Django
Django==1.10.2
django-model-utils==2.6
psycopg2==2.6.2
pytz==2016.7

# API
graphene==1.0.2
graphene-django==1.0
django-graph-auth==0.3.0
